+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## *Version 1.2.x breaking change will require the plugin config to be open and saved to display irrigation system.*
 
+## 1.2.12
+Update
+- Bumped dependencies.
+- Inital support for Homebridge v2.0.0
+-	Removed dependency on depratacted Homekit characteristic.
+- Code Cleanup
+
 ## 1.2.11
 Update
 - Bumped dependencies.